refactor(Balance): Delete old fragment (!!), fix imports
