+++ app/src/main/java/piuk/blockchain/android/data/connectivity/ConnectionStateMonitor.java
@@ -13,7 +13,7 @@ import android.support.v4.content.LocalBroadcastManager;
 import io.reactivex.Completable;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import piuk.blockchain.android.data.rxjava.IgnorableDefaultObserver;
-import piuk.blockchain.android.ui.balance.LegacyBalanceFragment;
+import piuk.blockchain.android.ui.balance.BalanceFragment;
 
 @TargetApi(Build.VERSION_CODES.LOLLIPOP)
 class ConnectionStateMonitor extends ConnectivityManager.NetworkCallback {
@@ -49,7 +49,7 @@ class ConnectionStateMonitor extends ConnectivityManager.NetworkCallback {
     private Completable broadcastOnMainThread() {
         return Completable.fromAction(() ->
                 LocalBroadcastManager.getInstance(context)
-                        .sendBroadcastSync(new Intent(LegacyBalanceFragment.ACTION_INTENT)))
+                        .sendBroadcastSync(new Intent(BalanceFragment.ACTION_INTENT)))
                 .subscribeOn(AndroidSchedulers.mainThread());
     }
