+++ app/src/main/java/piuk/blockchain/android/ui/backup/ConfirmFundsTransferDialogFragment.java
@@ -26,7 +26,7 @@ import uk.co.chrisjenx.calligraphy.TypefaceUtils;
 
 import piuk.blockchain.android.R;
 import piuk.blockchain.android.ui.account.SecondPasswordHandler;
-import piuk.blockchain.android.ui.balance.LegacyBalanceFragment;
+import piuk.blockchain.android.ui.balance.BalanceFragment;
 import piuk.blockchain.android.ui.customviews.MaterialProgressDialog;
 import piuk.blockchain.android.ui.customviews.ToastCustom;
 import piuk.blockchain.android.ui.send.AddressAdapter;
@@ -212,7 +212,7 @@ public class ConfirmFundsTransferDialogFragment extends AppCompatDialogFragment
 
     @Override
     public void dismissDialog() {
-        Intent intent = new Intent(LegacyBalanceFragment.ACTION_INTENT);
+        Intent intent = new Intent(BalanceFragment.ACTION_INTENT);
         LocalBroadcastManager.getInstance(getActivity()).sendBroadcast(intent);
         dismiss();
     }
