--- app/src/main/java/piuk/blockchain/android/ui/balance/BalanceViewModel.java
@@ -1,767 +1,0 @@
-package piuk.blockchain.android.ui.balance;
-
-import com.google.common.collect.HashBiMap;
-
-import android.support.annotation.NonNull;
-import android.support.annotation.StringRes;
-
-import info.blockchain.wallet.exceptions.ApiException;
-import info.blockchain.wallet.multiaddress.TransactionSummary;
-import info.blockchain.wallet.payload.data.Account;
-import info.blockchain.wallet.payload.data.LegacyAddress;
-
-import java.io.IOException;
-import java.math.BigInteger;
-import java.text.DecimalFormat;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-
-import javax.inject.Inject;
-
-import io.reactivex.Observable;
-import piuk.blockchain.android.R;
-import piuk.blockchain.android.data.access.AccessState;
-import piuk.blockchain.android.data.access.AuthEvent;
-import piuk.blockchain.android.data.contacts.ContactTransactionModel;
-import piuk.blockchain.android.data.contacts.ContactsEvent;
-import piuk.blockchain.android.data.datamanagers.BuyDataManager;
-import piuk.blockchain.android.data.datamanagers.ContactsDataManager;
-import piuk.blockchain.android.data.datamanagers.OnboardingDataManager;
-import piuk.blockchain.android.data.datamanagers.PayloadDataManager;
-import piuk.blockchain.android.data.datamanagers.TransactionListDataManager;
-import piuk.blockchain.android.data.notifications.NotificationPayload;
-import piuk.blockchain.android.data.rxjava.RxBus;
-import piuk.blockchain.android.data.settings.SettingsDataManager;
-import piuk.blockchain.android.injection.Injector;
-import piuk.blockchain.android.ui.account.ConsolidatedAccount;
-import piuk.blockchain.android.ui.account.ConsolidatedAccount.Type;
-import piuk.blockchain.android.ui.account.ItemAccount;
-import piuk.blockchain.android.ui.base.BaseViewModel;
-import piuk.blockchain.android.ui.customviews.ToastCustom;
-import piuk.blockchain.android.ui.home.MainActivity;
-import piuk.blockchain.android.ui.onboarding.OnboardingPagerContent;
-import piuk.blockchain.android.ui.swipetoreceive.SwipeToReceiveHelper;
-import piuk.blockchain.android.util.AppUtil;
-import piuk.blockchain.android.util.ExchangeRateFactory;
-import piuk.blockchain.android.util.MonetaryUtil;
-import piuk.blockchain.android.util.PrefsUtil;
-import piuk.blockchain.android.util.StringUtils;
-
-@SuppressWarnings("WeakerAccess")
-public class BalanceViewModel extends BaseViewModel {
-
-    private final String TAG = getClass().getName();
-
-    private DataListener dataListener;
-
-    private Observable<ContactsEvent> contactsEventObservable;
-    private Observable<NotificationPayload> notificationObservable;
-    private Observable<List> txListObservable;
-    private Observable<AuthEvent> authEventObservable;
-    private List<ItemAccount> activeAccountAndAddressList;
-    private HashBiMap<Object, Integer> activeAccountAndAddressBiMap;
-    private List<Object> displayList;
-    @Inject protected PrefsUtil prefsUtil;
-    @Inject protected StringUtils stringUtils;
-    @Inject protected TransactionListDataManager transactionListDataManager;
-    @Inject protected ContactsDataManager contactsDataManager;
-    @Inject protected PayloadDataManager payloadDataManager;
-    @Inject protected SettingsDataManager settingsDataManager;
-    @Inject protected SwipeToReceiveHelper swipeToReceiveHelper;
-    @Inject protected RxBus rxBus;
-    @Inject protected OnboardingDataManager onboardingDataManager;
-    @Inject protected ExchangeRateFactory exchangeRateFactory;
-    @Inject protected AppUtil appUtil;
-    @Inject protected BuyDataManager buyDataManager;
-    @Inject protected AccessState accessState;
-
-    public interface DataListener {
-
-        int getSelectedItemPosition();
-
-        boolean isBtc();
-
-        /**
-         * We can't simply call BuildConfig.CONTACTS_ENABLED in this class as it would make it
-         * impossible to test, as it's reliant on the build.gradle config. Passing it here
-         * allows us to change the response via mocking the DataListener.
-         *
-         * TODO: This should be removed once/if Contacts ships
-         */
-        boolean getIfContactsEnabled();
-
-        void onRefreshAccounts();
-
-        void onAccountSizeChange();
-
-        void onRefreshBalanceAndTransactions();
-
-        void onRefreshContactList();
-
-        void onPendingTxUpdate();
-
-        // TODO: 19/06/2017 moved to promptManager
-        void showBackupPromptDialog(boolean showNeverAgain);
-
-        // TODO: 19/06/2017 moved to promptManager
-        void show2FaDialog();
-
-        void updateBalance(String balance);
-
-        void setShowRefreshing(boolean showRefreshing);
-
-        void showToast(@StringRes int message, @ToastCustom.ToastType String toastType);
-
-        void showAccountChoiceDialog(List<String> accounts, String fctxId);
-
-        void initiatePayment(String uri, String recipientId, String mdid, String fctxId);
-
-        void showWaitingForPaymentDialog();
-
-        void showWaitingForAddressDialog();
-
-        void showSendAddressDialog(String fctxId);
-
-        void showProgressDialog();
-
-        void dismissProgressDialog();
-
-        void showFctxRequiringAttention(int number);
-
-        void showTransactionDeclineDialog(String fctxId);
-
-        void showTransactionCancelDialog(String fctxId);
-
-        void startBuyActivity();
-
-        void startReceiveFragment();
-
-        void onExchangeRateUpdated();
-
-        void onShowAnnouncement();
-
-        void onHideAnnouncement();
-
-        void onLoadOnboardingPages(List<OnboardingPagerContent> pages);
-    }
-
-    BalanceViewModel(DataListener dataListener) {
-        Injector.getInstance().getDataManagerComponent().inject(this);
-        this.dataListener = dataListener;
-
-        activeAccountAndAddressList = new ArrayList<>();
-        activeAccountAndAddressBiMap = HashBiMap.create();
-        displayList = new ArrayList<>();
-    }
-
-    @Override
-    public void onViewReady() {
-
-//        if(prefsUtil.getValue(PrefsUtil.KEY_APP_VISITS, 0) > 0) {
-//            // Check from this point forwards
-//            txListObservable = rxBus.register(List.class);
-//        }
-//
-//        compositeDisposable.add(
-//                exchangeRateFactory.updateTicker()
-//                        .subscribe(
-//                                () -> dataListener.onExchangeRateUpdated(),
-//                                Throwable::printStackTrace));
-//
-//        contactsEventObservable = rxBus.register(ContactsEvent.class);
-//        contactsEventObservable.subscribe(contactsEvent -> refreshFacilitatedTransactions());
-//
-//        authEventObservable = rxBus.register(AuthEvent.class);
-//        authEventObservable.subscribe(authEvent -> {
-//            displayList.clear();
-//            transactionListDataManager.clearTransactionList();
-//            contactsDataManager.resetContacts();
-//        });
-//
-//        notificationObservable = rxBus.register(NotificationPayload.class);
-//        notificationObservable
-//                .subscribe(notificationPayload -> {
-//                    if (notificationPayload.getType() != null
-//                            && notificationPayload.getType().equals(NotificationPayload.NotificationType.PAYMENT)) {
-//                        refreshFacilitatedTransactions();
-//                    }
-//                });
-    }
-
-    boolean areLauncherShortcutsEnabled() {
-        return prefsUtil.getValue(PrefsUtil.KEY_RECEIVE_SHORTCUTS_ENABLED, true);
-    }
-
-    public PayloadDataManager getPayloadDataManager() {
-        return payloadDataManager;
-    }
-
-    void storeSwipeReceiveAddresses() {
-        // Defer to background thread as deriving addresses is quite processor intensive
-//        compositeDisposable.add(
-//                Completable.fromCallable(() -> {
-//                    swipeToReceiveHelper.updateAndStoreAddresses();
-//                    return Void.TYPE;
-//                }).subscribeOn(Schedulers.computation())
-//                        .subscribe(() -> {
-//                            // No-op
-//                        }, Throwable::printStackTrace));
-    }
-
-    @SuppressWarnings("Convert2streamapi")
-    void updateAccountList() {
-        //activeAccountAndAddressList is linked to Adapter - do not reconstruct or loose reference otherwise notifyDataSetChanged won't work
-        activeAccountAndAddressList.clear();
-        activeAccountAndAddressBiMap.clear();
-
-        int spinnerIndex = 0;
-
-        //All accounts/addresses
-        List<Account> allAccounts;
-        List<LegacyAddress> allLegacyAddresses = payloadDataManager.getLegacyAddresses();
-
-        //Only active accounts/addresses (exclude archived)
-        List<Account> activeAccounts = new ArrayList<>();
-        allAccounts = payloadDataManager.getAccounts();//V3
-
-        for (Account item : allAccounts) {
-            if (!item.isArchived()) {
-                activeAccounts.add(item);
-            }
-        }
-
-        List<LegacyAddress> activeLegacyAddresses = new ArrayList<>();
-        for (LegacyAddress item : allLegacyAddresses) {
-            if (item.getTag() != LegacyAddress.ARCHIVED_ADDRESS) {
-                activeLegacyAddresses.add(item);
-            }
-        }
-
-        //"All" - total balance
-        if (activeAccounts.size() > 1 || !activeLegacyAddresses.isEmpty()) {
-            //Only V3 will display "All"
-            ConsolidatedAccount all = new ConsolidatedAccount();
-            all.setLabel(stringUtils.getString(R.string.all_accounts));
-            all.setType(Type.ALL_ACCOUNTS);
-
-            BigInteger bal = payloadDataManager.getWalletBalance();
-            String balance = getBalanceString(true, bal.longValue());
-            activeAccountAndAddressList.add(new ItemAccount(
-                    all.getLabel(),
-                    balance,
-                    null,
-                    bal.longValue(),
-                    null));
-            activeAccountAndAddressBiMap.put(all, spinnerIndex);
-            spinnerIndex++;
-        }
-
-        //Add accounts to map
-        int accountIndex = 0;
-        for (Account item : activeAccounts) {
-
-            //Give unlabeled account a label
-            if (item.getLabel().trim().isEmpty()) item.setLabel("Account: " + accountIndex);
-
-            BigInteger bal = payloadDataManager.getAddressBalance(item.getXpub());
-            String balanceString = getBalanceString(true, bal.longValue());
-
-            activeAccountAndAddressList.add(new ItemAccount(
-                    item.getLabel(),
-                    balanceString,
-                    null,
-                    bal.longValue(),
-                    null));
-            activeAccountAndAddressBiMap.put(item, spinnerIndex);
-            spinnerIndex++;
-            accountIndex++;
-        }
-
-        //Add "Imported Addresses" or "Total Funds" to map
-        if (!activeLegacyAddresses.isEmpty()) {
-            //Only V3 - Consolidate and add Legacy addresses to "Imported Addresses" at bottom of accounts spinner
-
-            ConsolidatedAccount importedAddresses = new ConsolidatedAccount();
-            importedAddresses.setLabel(stringUtils.getString(R.string.imported_addresses));
-            importedAddresses.setType(Type.ALL_IMPORTED_ADDRESSES);
-
-            BigInteger bal = payloadDataManager.getImportedAddressesBalance();
-            String balance = getBalanceString(true, bal.longValue());
-
-            activeAccountAndAddressList.add(new ItemAccount(
-                    importedAddresses.getLabel(),
-                    balance,
-                    null,
-                    bal.longValue(),
-                    null));
-            activeAccountAndAddressBiMap.put(importedAddresses, spinnerIndex);
-        }
-
-        //If we have multiple accounts/addresses we will show dropdown in toolbar, otherwise we will only display a static text
-        if (dataListener != null) {
-            dataListener.onRefreshAccounts();
-        }
-    }
-
-    List<Object> getTransactionList() {
-        return displayList;
-    }
-
-    HashMap<String, String> getContactsTransactionMap() {
-        return contactsDataManager.getContactsTransactionMap();
-    }
-
-    HashMap<String, String> getNotesTransactionMap() {
-        return contactsDataManager.getNotesTransactionMap();
-    }
-
-    double getLastPrice(String fiat) {
-        return exchangeRateFactory.getLastPrice(fiat);
-    }
-
-    void onTransactionListRefreshed() {
-//        compositeDisposable.add(
-//                payloadDataManager.updateAllTransactions()
-//                        .doOnSubscribe(disposable -> dataListener.setShowRefreshing(true))
-//                        .doAfterTerminate(() -> dataListener.setShowRefreshing(false))
-//                        .subscribe(() -> {
-//                            updateAccountList();
-//                            refreshFacilitatedTransactions();
-//                            updateBalanceAndTransactionList(dataListener.getSelectedItemPosition(), dataListener.isBtc(), true);
-//                        }, throwable -> {
-//                            // No-op
-//                        }));
-    }
-
-    void updateBalanceAndTransactionList(int accountSpinnerPosition, boolean isBTC, boolean fetchTransactions) {
-//        // The current selected item in dropdown (Account or Legacy Address)
-//        Object object = activeAccountAndAddressBiMap.inverse().get(accountSpinnerPosition);
-//
-//        //If current selected item gets edited by another platform object might become null
-//        if (object == null && dataListener != null) {
-//            dataListener.onAccountSizeChange();
-//            object = activeAccountAndAddressBiMap.inverse().get(accountSpinnerPosition);
-//        }
-//
-//        //Update balance
-//        final Object finalObject = object;
-//        compositeDisposable.add(
-//                payloadDataManager.updateAllBalances()
-//                        .subscribe(() -> {
-//                            long btcBalance = transactionListDataManager.getBtcBalance(finalObject);
-//                            String balanceTotal = getBalanceString(isBTC, btcBalance);
-//
-//                            if (dataListener != null) {
-//                                dataListener.updateBalance(balanceTotal);
-//                            }
-//                        }));
-//
-//        compositeDisposable.add(
-//                Observable.just(fetchTransactions)
-//                        .flatMap(fetch -> {
-//                            if (fetch) {
-//                                return transactionListDataManager.fetchTransactions(finalObject, 50, 0);
-//                            } else {
-//                                return Observable.just(transactionListDataManager.getTransactionList());
-//                            }
-//                        })
-//                        .doAfterTerminate(() -> {
-//                            if (dataListener != null) dataListener.setShowRefreshing(false);
-//                        })
-//                        .subscribe(
-//                                this::insertTransactionsAndDisplay,
-//                                throwable -> Log.e(TAG, "updateBalanceAndTransactionList: ", throwable)));
-    }
-
-    @SuppressWarnings("Java8CollectionRemoveIf")
-    private void insertTransactionsAndDisplay(List<TransactionSummary> txList) throws IOException, ApiException {
-        // Remove current transactions but keep headers and pending transactions
-//        Iterator iterator = displayList.iterator();
-//        while (iterator.hasNext()) {
-//            Object element = iterator.next();
-//            if (element instanceof TransactionSummary) {
-//                iterator.remove();
-//            }
-//        }
-//
-//        displayList.addAll(txList);
-//        checkLatestAnnouncement(txList);
-//
-//        if (dataListener != null) {
-//            dataListener.onRefreshBalanceAndTransactions();
-//        }
-    }
-
-    void refreshFacilitatedTransactions() {
-//        if (dataListener.getIfContactsEnabled()) {
-//            compositeDisposable.add(
-//                    contactsDataManager.fetchContacts()
-//                            .andThen(contactsDataManager.getContactsWithUnreadPaymentRequests())
-//                            .toList()
-//                            .doOnSuccess(contacts -> dataListener.onRefreshContactList())
-//                            .flatMapObservable(contacts -> contactsDataManager.refreshFacilitatedTransactions())
-//                            .toList()
-//                            .subscribe(
-//                                    this::handlePendingTransactions,
-//                                    Throwable::printStackTrace));
-//        }
-    }
-
-    void getFacilitatedTransactions() {
-//        compositeDisposable.add(
-//                contactsDataManager.getFacilitatedTransactions()
-//                        .toList()
-//                        .doOnSuccess(contacts -> dataListener.onRefreshContactList())
-//                        .subscribe(
-//                                this::handlePendingTransactions,
-//                                Throwable::printStackTrace));
-    }
-
-    void onPendingTransactionClicked(String fctxId) {
-//        compositeDisposable.add(
-//                contactsDataManager.getContactFromFctxId(fctxId)
-//                        .subscribe(contact -> {
-//                            FacilitatedTransaction transaction = contact.getFacilitatedTransactions().get(fctxId);
-//
-//                            if (transaction == null) {
-//                                dataListener.showToast(R.string.contacts_transaction_not_found_error, ToastCustom.TYPE_ERROR);
-//                            } else {
-//
-//                                // Payment request sent, waiting for address from recipient
-//                                if (transaction.getState().equals(FacilitatedTransaction.STATE_WAITING_FOR_ADDRESS)
-//                                        && transaction.getRole().equals(FacilitatedTransaction.ROLE_RPR_INITIATOR)) {
-//
-//                                    dataListener.showWaitingForAddressDialog();
-//
-//                                    // Payment request sent, waiting for payment
-//                                } else if (transaction.getState().equals(FacilitatedTransaction.STATE_WAITING_FOR_PAYMENT)
-//                                        && transaction.getRole().equals(FacilitatedTransaction.ROLE_PR_INITIATOR)) {
-//
-//                                    dataListener.showWaitingForPaymentDialog();
-//
-//                                    // Received payment request, need to send address to sender
-//                                } else if (transaction.getState().equals(FacilitatedTransaction.STATE_WAITING_FOR_ADDRESS)
-//                                        && transaction.getRole().equals(FacilitatedTransaction.ROLE_PR_RECEIVER)) {
-//
-//                                    List<String> accountNames = new ArrayList<>();
-//                                    //noinspection Convert2streamapi
-//                                    for (Account account : payloadDataManager.getAccounts()) {
-//                                        if (!account.isArchived()) {
-//                                            accountNames.add(account.getLabel());
-//                                        }
-//                                    }
-//
-//                                    if (accountNames.size() == 1) {
-//                                        // Only one account, ask if you want to send an address
-//                                        dataListener.showSendAddressDialog(fctxId);
-//                                    } else {
-//                                        // Show dialog allowing user to select which account they want to use
-//                                        dataListener.showAccountChoiceDialog(accountNames, fctxId);
-//                                    }
-//
-//                                    // Waiting for payment
-//                                } else if (transaction.getState().equals(FacilitatedTransaction.STATE_WAITING_FOR_PAYMENT)
-//                                        && transaction.getRole().equals(FacilitatedTransaction.ROLE_RPR_RECEIVER)) {
-//
-//                                    dataListener.initiatePayment(
-//                                            transaction.toBitcoinURI(),
-//                                            contact.getId(),
-//                                            contact.getMdid(),
-//                                            transaction.getId());
-//                                }
-//                            }
-//                        }, throwable -> dataListener.showToast(R.string.contacts_transaction_not_found_error, ToastCustom.TYPE_ERROR)));
-    }
-
-    void onPendingTransactionLongClicked(String fctxId) {
-//        compositeDisposable.add(
-//                contactsDataManager.getFacilitatedTransactions()
-//                        .filter(contactTransactionModel -> contactTransactionModel.getFacilitatedTransaction().getId().equals(fctxId))
-//                        .subscribe(contactTransactionModel -> {
-//                            FacilitatedTransaction fctx = contactTransactionModel.getFacilitatedTransaction();
-//
-//                            if (fctx.getState().equals(FacilitatedTransaction.STATE_WAITING_FOR_ADDRESS)) {
-//                                if (fctx.getRole().equals(FacilitatedTransaction.ROLE_PR_RECEIVER)) {
-//                                    dataListener.showTransactionDeclineDialog(fctxId);
-//                                } else if (fctx.getRole().equals(FacilitatedTransaction.ROLE_RPR_INITIATOR)) {
-//                                    dataListener.showTransactionCancelDialog(fctxId);
-//                                }
-//
-//                            } else if (fctx.getState().equals(FacilitatedTransaction.STATE_WAITING_FOR_PAYMENT)) {
-//                                if (fctx.getRole().equals(FacilitatedTransaction.ROLE_RPR_RECEIVER)) {
-//                                    dataListener.showTransactionDeclineDialog(fctxId);
-//                                } else if (fctx.getRole().equals(FacilitatedTransaction.ROLE_PR_INITIATOR)) {
-//                                    dataListener.showTransactionCancelDialog(fctxId);
-//                                }
-//                            }
-//                        }, throwable -> {
-//                            // No-op
-//                        }));
-    }
-
-    void confirmDeclineTransaction(String fctxId) {
-//        compositeDisposable.add(
-//                contactsDataManager.getContactFromFctxId(fctxId)
-//                        .flatMapCompletable(contact -> contactsDataManager.sendPaymentDeclinedResponse(contact.getMdid(), fctxId))
-//                        .doOnError(throwable -> contactsDataManager.fetchContacts())
-//                        .doAfterTerminate(this::refreshFacilitatedTransactions)
-//                        .subscribe(
-//                                () -> dataListener.showToast(R.string.contacts_pending_transaction_decline_success, ToastCustom.TYPE_OK),
-//                                throwable -> dataListener.showToast(R.string.contacts_pending_transaction_decline_failure, ToastCustom.TYPE_ERROR)));
-    }
-
-    void confirmCancelTransaction(String fctxId) {
-//        compositeDisposable.add(
-//                contactsDataManager.getContactFromFctxId(fctxId)
-//                        .flatMapCompletable(contact -> contactsDataManager.sendPaymentCancelledResponse(contact.getMdid(), fctxId))
-//                        .doOnError(throwable -> contactsDataManager.fetchContacts())
-//                        .doAfterTerminate(this::refreshFacilitatedTransactions)
-//                        .subscribe(
-//                                () -> dataListener.showToast(R.string.contacts_pending_transaction_cancel_success, ToastCustom.TYPE_OK),
-//                                throwable -> dataListener.showToast(R.string.contacts_pending_transaction_cancel_failure, ToastCustom.TYPE_ERROR)));
-    }
-
-    void onAccountChosen(int accountPosition, String fctxId) {
-//        dataListener.showProgressDialog();
-//        compositeDisposable.add(
-//                contactsDataManager.getContactFromFctxId(fctxId)
-//                        .subscribe(contact -> {
-//                            FacilitatedTransaction transaction = contact.getFacilitatedTransactions().get(fctxId);
-//
-//                            PaymentRequest paymentRequest = new PaymentRequest();
-//                            paymentRequest.setIntendedAmount(transaction.getIntendedAmount());
-//                            paymentRequest.setId(fctxId);
-//
-//                            compositeDisposable.add(
-//                                    payloadDataManager.getNextReceiveAddressAndReserve(getCorrectedAccountIndex(accountPosition), "Payment request " + transaction.getId())
-//                                            .doOnNext(paymentRequest::setAddress)
-//                                            .flatMapCompletable(s -> contactsDataManager.sendPaymentRequestResponse(contact.getMdid(), paymentRequest, fctxId))
-//                                            .doAfterTerminate(() -> dataListener.dismissProgressDialog())
-//                                            .subscribe(
-//                                                    () -> {
-//                                                        dataListener.showToast(R.string.contacts_address_sent_success, ToastCustom.TYPE_OK);
-//                                                        refreshFacilitatedTransactions();
-//                                                    },
-//                                                    throwable -> dataListener.showToast(R.string.contacts_address_sent_failed, ToastCustom.TYPE_ERROR)));
-//                        }, throwable -> dataListener.showToast(R.string.contacts_transaction_not_found_error, ToastCustom.TYPE_ERROR)));
-    }
-
-    @NonNull
-    private String getBalanceString(boolean isBTC, long btcBalance) {
-//        double fiatBalance;
-//        String strFiat = prefsUtil.getValue(PrefsUtil.KEY_SELECTED_FIAT, PrefsUtil.DEFAULT_CURRENCY);
-//        double lastPrice = ExchangeRateFactory.getInstance().getLastPrice(strFiat);
-//        fiatBalance = lastPrice * (btcBalance / 1e8);
-//
-//        return isBTC ? getMonetaryUtil().getDisplayAmountWithFormatting(btcBalance) + " " + getDisplayUnits()
-//                : getMonetaryUtil().getFiatFormat(strFiat).format(fiatBalance) + " " + strFiat;
-        return "";
-    }
-
-    public String getDisplayUnits() {
-        return (String) getMonetaryUtil().getBTCUnits()[prefsUtil.getValue(PrefsUtil.KEY_BTC_UNITS, MonetaryUtil.UNIT_BTC)];
-    }
-
-    /**
-     * Saves value of true to prevent users from seeing the backup prompt again
-     */
-    // TODO: 19/06/2017 moved to promptManager
-    public void neverPromptBackup() {
-        prefsUtil.setValue(PrefsUtil.KEY_SECURITY_BACKUP_NEVER, true);
-    }
-
-    /**
-     * Saves value of true to prevent users from seeing the 2FA prompt again
-     */
-    // TODO: 19/06/2017 moved to promptManager
-    public void neverPrompt2Fa() {
-        prefsUtil.setValue(PrefsUtil.KEY_SECURITY_TWO_FA_NEVER, true);
-    }
-
-    public List<ItemAccount> getActiveAccountAndAddressList() {
-        return activeAccountAndAddressList;
-    }
-
-    public List<Account> getAccounts() {
-        return payloadDataManager.getAccounts();
-    }
-
-    public MonetaryUtil getMonetaryUtil() {
-        return new MonetaryUtil(prefsUtil.getValue(PrefsUtil.KEY_BTC_UNITS, MonetaryUtil.UNIT_BTC));
-    }
-
-    public PrefsUtil getPrefsUtil() {
-        return prefsUtil;
-    }
-
-    // TODO: 19/06/2017 moved to promptManager
-    private boolean hasTransactions() {
-        return !transactionListDataManager.getTransactionList().isEmpty();
-    }
-
-    private int getCorrectedAccountIndex(int accountIndex) {
-//        // Filter accounts by active
-//        List<Account> activeAccounts = new ArrayList<>();
-//        List<Account> accounts = payloadDataManager.getAccounts();
-//        for (int i = 0; i < accounts.size(); i++) {
-//            Account account = accounts.get(i);
-//            if (!account.isArchived()) {
-//                activeAccounts.add(account);
-//            }
-//        }
-//
-//        // Find corrected position
-//        return payloadDataManager.getAccounts().indexOf(activeAccounts.get(accountIndex));
-        return -1;
-    }
-
-    private int getNumberOfFctxRequiringAttention(List<ContactTransactionModel> facilitatedTransactions) {
-//        int value = 0;
-//        for (ContactTransactionModel transactionModel : facilitatedTransactions) {
-//            FacilitatedTransaction transaction = transactionModel.getFacilitatedTransaction();
-//            if (transaction.getState().equals(FacilitatedTransaction.STATE_WAITING_FOR_ADDRESS)
-//                    && transaction.getRole().equals(FacilitatedTransaction.ROLE_RPR_RECEIVER)) {
-//                value++;
-//            } else if (transaction.getState().equals(FacilitatedTransaction.STATE_WAITING_FOR_PAYMENT)
-//                    && transaction.getRole().equals(FacilitatedTransaction.ROLE_RPR_RECEIVER)) {
-//                value++;
-//            }
-//        }
-//        return value;
-        return -1;
-    }
-
-    @SuppressWarnings("Java8CollectionRemoveIf")
-    private void handlePendingTransactions(List<ContactTransactionModel> transactions) {
-        // Remove previous Pending Transactions
-//        Iterator iterator = displayList.iterator();
-//        while (iterator.hasNext()) {
-//            Object element = iterator.next();
-//            if (!(element instanceof TransactionSummary)) {
-//                iterator.remove();
-//            }
-//        }
-//
-//        dataListener.showFctxRequiringAttention(getNumberOfFctxRequiringAttention(transactions));
-//
-//        if (!transactions.isEmpty()) {
-//            //noinspection Java8ListSort
-//            Collections.sort(transactions, new ContactTransactionDateComparator());
-//            Collections.reverse(transactions);
-//            displayList.add(0, stringUtils.getString(R.string.contacts_pending_transaction));
-//            displayList.addAll(1, transactions);
-//            displayList.add(transactions.size() + 1, stringUtils.getString(R.string.contacts_transaction_history));
-//            dataListener.onPendingTxUpdate();
-//        } else {
-//            dataListener.onPendingTxUpdate();
-//        }
-    }
-
-    @Override
-    public void destroy() {
-//        rxBus.unregister(ContactsEvent.class, contactsEventObservable);
-//        rxBus.unregister(NotificationPayload.class, notificationObservable);
-//        rxBus.unregister(List.class, txListObservable);
-//        rxBus.unregister(AuthEvent.class, authEventObservable);
-        super.destroy();
-    }
-
-    // TODO: 21/06/2017
-    void addOnboardingPages() {
-        compositeDisposable.add(
-                buyDataManager.getCanBuy()
-                        .subscribe(isBuyAllowed -> {
-                            List<OnboardingPagerContent> onboardingPages = getOnboardingPages(isBuyAllowed);
-                            dataListener.onLoadOnboardingPages(onboardingPages);
-                        }, Throwable::printStackTrace));
-    }
-
-    private List<OnboardingPagerContent> getOnboardingPages(boolean isBuyAllowed) {
-
-        List<OnboardingPagerContent> pages = new ArrayList<>();
-        if (isBuyAllowed) {
-            pages.add(new OnboardingPagerContent(stringUtils.getString(R.string.onboarding_current_price),
-                    getFormattedPriceString(),
-                    stringUtils.getString(R.string.onboarding_buy_content),
-                    stringUtils.getString(R.string.onboarding_buy_bitcoin),
-                    MainActivity.ACTION_BUY,
-                    R.color.primary_blue_accent,
-                    R.drawable.vector_buy_offset));
-        }
-
-        //Receive bitcoin
-        pages.add(new OnboardingPagerContent(stringUtils.getString(R.string.onboarding_receive_bitcoin),
-                "",
-                stringUtils.getString(R.string.onboarding_receive_content),
-                stringUtils.getString(R.string.receive_bitcoin),
-                MainActivity.ACTION_RECEIVE,
-                R.color.secondary_teal_medium,
-                R.drawable.vector_receive_offset));
-
-        //QR Codes
-        pages.add(new OnboardingPagerContent(stringUtils.getString(R.string.onboarding_qr_codes),
-                "",
-                stringUtils.getString(R.string.onboarding_qr_codes_content),
-                stringUtils.getString(R.string.onboarding_scan_address),
-                MainActivity.ACTION_SEND,
-                R.color.primary_navy_medium,
-                R.drawable.vector_qr_offset));
-        return pages;
-    }
-
-    private String getFormattedPriceString() {
-        String fiat = prefsUtil.getValue(PrefsUtil.KEY_SELECTED_FIAT, "");
-        double lastPrice = exchangeRateFactory.getLastPrice(fiat);
-        String fiatSymbol = exchangeRateFactory.getSymbol(fiat);
-        DecimalFormat format = new DecimalFormat();
-        format.setMinimumFractionDigits(2);
-        return stringUtils.getFormattedString(
-                R.string.current_price_btc,
-                fiatSymbol + format.format(lastPrice));
-    }
-
-    public boolean isOnboardingComplete() {
-        // If wallet isn't newly created, don't show onboarding
-        return prefsUtil.getValue(PrefsUtil.KEY_ONBOARDING_COMPLETE, false)
-                || !appUtil.isNewlyCreated();
-    }
-
-    public void setOnboardingComplete(boolean completed) {
-        prefsUtil.setValue(PrefsUtil.KEY_ONBOARDING_COMPLETE, completed);
-    }
-
-    public void getBitcoinClicked() {
-        compositeDisposable.add(
-                buyDataManager.getCanBuy()
-                        .subscribe(buyAllowed -> {
-                            if (buyAllowed) {
-                                dataListener.startBuyActivity();
-                            } else {
-                                dataListener.startReceiveFragment();
-                            }
-                        }, Throwable::printStackTrace));
-    }
-
-    public void checkLatestAnnouncement(List<TransactionSummary> txList) {
-        //If user hasn't completed onboarding, ignore announcements
-        compositeDisposable.add(
-                buyDataManager.getCanBuy()
-                        .subscribe(buyAllowed -> {
-                            if (isOnboardingComplete() && buyAllowed) {
-                                if (!prefsUtil.getValue(PrefsUtil.KEY_LATEST_ANNOUNCEMENT_DISMISSED, false) && !txList.isEmpty()) {
-                                    prefsUtil.setValue(PrefsUtil.KEY_LATEST_ANNOUNCEMENT_SEEN, true);
-                                    dataListener.onShowAnnouncement();
-                                } else {
-                                    dataListener.onHideAnnouncement();
-                                }
-                            } else {
-                                dataListener.onHideAnnouncement();
-                            }
-                        }));
-    }
-
-    public void disableAnnouncement() {
-        prefsUtil.setValue(PrefsUtil.KEY_LATEST_ANNOUNCEMENT_DISMISSED, true);
-    }
-}
