+++ app/src/main/java/piuk/blockchain/android/ui/contacts/detail/ContactDetailFragment.java
@@ -33,8 +33,8 @@ import piuk.blockchain.android.util.PrefsUtil;
 import piuk.blockchain.android.util.ViewUtils;
 import piuk.blockchain.android.util.annotations.Thunk;
 
-import static piuk.blockchain.android.ui.balance.LegacyBalanceFragment.SHOW_BTC;
-import static piuk.blockchain.android.ui.send.SendViewModel.SHOW_FIAT;
+import static piuk.blockchain.android.data.access.AccessState.SHOW_BTC;
+import static piuk.blockchain.android.data.access.AccessState.SHOW_FIAT;
 
 
 public class ContactDetailFragment extends Fragment implements ContactDetailViewModel.DataListener {
