+++ app/src/main/java/piuk/blockchain/android/ui/send/SendFragment.java
@@ -68,7 +68,7 @@ import piuk.blockchain.android.databinding.FragmentSendBinding;
 import piuk.blockchain.android.ui.account.ItemAccount;
 import piuk.blockchain.android.ui.account.PaymentConfirmationDetails;
 import piuk.blockchain.android.ui.account.SecondPasswordHandler;
-import piuk.blockchain.android.ui.balance.LegacyBalanceFragment;
+import piuk.blockchain.android.ui.balance.BalanceFragment;
 import piuk.blockchain.android.ui.base.BaseAuthActivity;
 import piuk.blockchain.android.ui.chooser.AccountChooserActivity;
 import piuk.blockchain.android.ui.confirm.ConfirmPaymentDialog;
@@ -128,7 +128,7 @@ public class SendFragment extends Fragment implements SendContract.DataListener,
     private BroadcastReceiver receiver = new BroadcastReceiver() {
         @Override
         public void onReceive(final Context context, final Intent intent) {
-            if (intent.getAction().equals(LegacyBalanceFragment.ACTION_INTENT) && binding != null) {
+            if (intent.getAction().equals(BalanceFragment.ACTION_INTENT) && binding != null) {
                 viewModel.updateUI();
             }
         }
@@ -185,7 +185,7 @@ public class SendFragment extends Fragment implements SendContract.DataListener,
         super.onResume();
         setupToolbar();
         closeKeypad();
-        IntentFilter filter = new IntentFilter(LegacyBalanceFragment.ACTION_INTENT);
+        IntentFilter filter = new IntentFilter(BalanceFragment.ACTION_INTENT);
         LocalBroadcastManager.getInstance(getActivity()).registerReceiver(receiver, filter);
         viewModel.updateUI();
     }
